Add script which loads only wanted data from MongoDB into CSV files.
For each graph/query is made a bunch of needed files with prefix like A1 or B2 etc.
Column selection takes less disk space. For other hand, there are partly common files, where we could use the one with more information. It will lead to more chaotic usage of these files as we would not know which we need and so on.